// types.ts
export type PropertyType = {
  _id?: string;
  owner?: string;
  name: string;
  type: string;
  description: string;
  location: {
    street: string;
    city: string;
    state: string;
    zipcode: string;
  };
  beds: number;
  baths: number;
  square_feet: number;
  amenities: string[];
  rates: {
    nightly?: number;
    weekly?: number;
    monthly?: number;
  };
  seller_info: {
    name: string;
    email: string;
    phone: string;
  };
  images: string[];
  is_featured?: boolean;
  createdAt?: string;
  updatedAt?: string;
};

export type UserType = {
  _id?: string;
  email: string;
  username: string;
  image?: string;
  bookmarks?: [PropertyType["_id"]];
};

export type MessageType = {
  _id?: string;
  sender?: UserType["_id"];
  recipient?: UserType["_id"];
  property?: PropertyType["_id"] | PropertyType;
  name: string;
  email: string;
  phone: string;
  message: string;
  read?: boolean;
  createdAt?: string;
  updatedAt?: string;
};